Remove mockedFunction for Node.js test runner mock method
